# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MusicDrop",
    "version": "dev"
  },
  "paths": {
...
    "/api/artists": {
      "get": {
        "tags": [
          "artists"
        ],
        "summary": "List Artists Endpoint",
        "operationId": "list_artists_endpoint_api_artists_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Artist"
                  },
                  "title": "Response List Artists Endpoint Api Artists Get"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "artists"
        ],
        "summary": "Delete Artist Endpoint",
        "description": "Move EVERY album of the named artist to Trash (reversible) and drop them.\n\n``name`` is a query param so slashes (e.g. \"AC/DC\") survive routing. 409\nwhile a library job is running.",
        "operationId": "delete_artist_endpoint_api_artists_delete",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/artists/image": {
      "get": {
        "tags": [
          "artists"
        ],
        "summary": "Get Artist Image Endpoint",
        "operationId": "get_artist_image_endpoint_api_artists_image_get",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The artist portrait.",
            "content": {
              "application/json": {
                "schema": {}
              },
              "image/*": {}
            }
          },
          "404": {
            "description": "Feature disabled, no verified match, or transient error."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/artists/image/settings": {
      "get": {
        "tags": [
          "artists"
        ],
        "summary": "Get Artist Image Settings Endpoint",
        "operationId": "get_artist_image_settings_endpoint_api_artists_image_settings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtistImageSettings"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "artists"
        ],
        "summary": "Set Artist Image Settings Endpoint",
        "operationId": "set_artist_image_settings_endpoint_api_artists_image_settings_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArtistImageSettings"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtistImageSettings"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/artists/image/override": {
      "post": {
        "tags": [
          "artists"
        ],
        "summary": "Upload Artist Image Override Endpoint",
        "operationId": "upload_artist_image_override_endpoint_api_artists_image_override_post",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_artist_image_override_endpoint_api_artists_image_override_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtistImageOverrideResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "artists"
        ],
        "summary": "Clear Artist Image Override Endpoint",
        "operationId": "clear_artist_image_override_endpoint_api_artists_image_override_delete",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "Name"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/artists/image/override/from-url": {
      "post": {
        "tags": [
          "artists"
        ],
        "summary": "Set Artist Image Override From Url Endpoint",
        "operationId": "set_artist_image_override_from_url_endpoint_api_artists_image_override_from_url_post",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArtistImageUrlOverride"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtistImageOverrideResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/artists/art/settings": {
      "get": {
        "tags": [
          "artists"
        ],
        "summary": "Get Artist Art Settings",
        "operationId": "get_artist_art_settings_api_artists_art_settings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtistArtWriteSettings"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "artists"
        ],
        "summary": "Set Artist Art Settings",
        "operationId": "set_artist_art_settings_api_artists_art_settings_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArtistArtWriteSettings"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtistArtWriteSettings"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/artists/art/apply": {
      "post": {
        "tags": [
          "artists"
        ],
        "summary": "Apply Artist Art",
        "operationId": "apply_artist_art_api_artists_art_apply_post",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtistArtBackfillStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Artist": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "album_count": {
            "type": "integer",
            "title": "Album Count"
          }
        },
        "type": "object",
        "required": [
          "name",
          "album_count"
        ],
        "title": "Artist"
      },
      "ArtistArtBackfillStatus": {
        "properties": {
          "phase": {
            "type": "string",
            "enum": [
              "idle",
              "running",
              "done",
              "stopped",
              "failed"
            ],
            "title": "Phase"
          },
          "job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "processed": {
            "type": "integer",
            "title": "Processed"
          },
          "written": {
            "type": "integer",
            "title": "Written"
          },
          "skipped": {
            "type": "integer",
            "title": "Skipped"
          },
          "failed": {
            "type": "integer",
            "title": "Failed"
          },
          "current": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "artist": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artist"
          },
          "scope_label": {
            "type": "string",
            "title": "Scope Label"
          }
        },
        "type": "object",
        "required": [
          "phase",
          "job_id",
          "total",
          "processed",
          "written",
          "skipped",
          "failed",
          "current",
          "error",
          "artist",
          "scope_label"
        ],
        "title": "ArtistArtBackfillStatus"
      },
      "ArtistArtWriteSettings": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          }
        },
        "type": "object",
        "required": [
          "enabled"
        ],
        "title": "ArtistArtWriteSettings"
      },
      "ArtistImageOverrideResult": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "content_type": {
            "type": "string",
            "title": "Content Type"
          }
        },
        "type": "object",
        "required": [
          "ok",
          "content_type"
        ],
        "title": "ArtistImageOverrideResult",
        "description": "Outcome of pinning a manual artist-image override."
      },
      "ArtistImageSettings": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          }
        },
        "type": "object",
        "required": [
          "enabled"
        ],
        "title": "ArtistImageSettings",
        "description": "The artist-image feature on/off flag (GET + PUT body + PUT response)."
      },
      "ArtistImageUrlOverride": {
        "properties": {
          "url": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "ArtistImageUrlOverride",
        "description": "Request body for setting an artist override from an image URL."
      },
...
      "Body_upload_artist_image_override_endpoint_api_artists_image_override_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_artist_image_override_endpoint_api_artists_image_override_post"
      },
...
      "DeleteResult": {
        "properties": {
          "trashed_albums": {
            "type": "integer",
            "title": "Trashed Albums"
          },
          "trash_path": {
            "type": "string",
            "title": "Trash Path"
          }
        },
        "type": "object",
        "required": [
          "trashed_albums",
          "trash_path"
        ],
        "title": "DeleteResult",
        "description": "Outcome of a reversible delete: how many albums went to Trash + where.\n\n``trashed_albums`` is 1 for a single-album delete, N for an artist (every\nalbum of theirs). ``trash_path`` is the Trash location the files were moved\nto (recoverable from there)."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}